+++ tests/test_graphql.py
@@ -71,6 +71,11 @@ def test_invalid_field():
         ],
     }
 
+def test_graphiql_get():
+    response = client.get("/", headers={"accept": "text/html"})
+    assert response.status_code == 200
+    assert "<!DOCTYPE html>" in response.text
+
 
 class AsyncQuery(graphene.ObjectType):
     hello = graphene.String(name=graphene.String(default_value='stranger'))